FPVTL-953: added Barrister DTO
